+++ projects/ui-lib/src/lib/modules/icons/icon.module.ts
@@ -1,10 +1,11 @@
 import { NgModule } from '@angular/core';
-import { ArrowLeft, ArrowRight, ArrowUpRight, LucideAngularModule } from 'lucide-angular';
+import { ArrowLeft, ArrowRight, ArrowUpRight, LucideAngularModule, Twitter } from 'lucide-angular';
 
 const icons = {
     ArrowUpRight,
     ArrowRight,
     ArrowLeft,
+    Twitter
 };
 
 @NgModule({